Implements time compensation for recordings
Sometimes it is neccessary to compensate for time offset or drift for
recording files. This implements a class which can use the known and
recorded time for events to compensate this offset and drift, typically
by storing an alternate start time for the files that are read.

This works fine for small drifts, where the drift within each file is
negligable.