Changes Branch to proper mapping
Branches are now mutable mappings, storing each child with a name
unrelated to it's metadata. This shold simplify the implementation
of stacking operations since it's not essential that the names of the
children live as metadata with the same name as the layer.
The layer attribute is also renamed to `dim`, inspired by xarray.
This is now a mandatory public argument.