Big rehaul for use xarray
This starts moving a lot of the handling of data over to the xarray
library.
This change touches almost every piece of code written, so the xarray
library will now be fully integrated into the package.

Also included is some of the function earlier used for analysis in
notebooks, like background compensation.