New recorder class
It will be easier to extend in the future if we split the recording
classes and the specification of where sensors are located.
This intoduces a very simple structure to deal with sensor locations,
and implements a new top level class for recorders. This will replace
the old hydrophone class, which tried to be too much at once.