Changes time_data to normal method
When composing objects on top of the recorders, it can be useful to be
able to assign the actual time_data method without calling it. This is
much easier to do if it's not a property.
It will also give better symmetry in the future if we add more data
methods that both read and process the recorded data.