Fixes data reading without sensor information
Sometimes the sensor information is unknown or not neccessary. In those
cases it is still useful to be able to read data using the recording classes.
These changes allows the recordings to be used even when no sensor
information is given, but the read time data will have a 'channel'
dimension instead of a 'sensor' dimension.