Separates base class for single and multiple positions
There's certaint things that just doesn't make sense for Line/Track to
have, but Position needs them. This shows that we need a dedicated base
class to handle coordinates generally.
This base class implements the mapping protocol, allowing data to be
stored in the underlieing xarray via indexing the object.