Adds attribute transfer to data wrapper
For some subclasses it makes sense to transfer some attributes that are
not stored in the core data when performing operations to create new
objects.
The `_transfer_attributes` function can be extended by subclasses to
move those attributes to new instances, and used by processing functions
to not loose those attributes.