Fixes init order for data wrappers
We were using a very strange initialization order for data wrappers,
where each subclass did work to prepare itsef, then the data, then passed
it upwards in the class hierarchy.
The new structure is that the base class does some work to make sure
that there is a data object to work with, then each subclass will modify
it to add the relevant stuff.
This will likely be easier to subclass in the future, since you can
use the superclass init easier.