Properly wraps more numpy functionality
This properly wraps almost all "simple" numpy functions. Most of the
lifting is done either by __array_ufunc__ which handles all ufuncs.
The reminder of the numpy api goes through __array_funciton__ which uses
a few of the methods defined in the wrapper, but mostly passes the arguments
through as xarrays and hopes for the best.