Implemetns background compensation
This background compensation takes frequency data as the input.
The call method removes the background noise from a received spectrum.
If the background noise is not using the same frequency axis as
the received spectrum, the background noise will be linearly interpolated.
This works for unequal bandwidths as long as both object are expressed
as a power spectral density.