Implements rolling time windows
This has two parts: one on the basic file recorder and one on audio file
recorder. The basic file recorder is responsible for reading the frames,
but outputs uncalibrated raw data.
The audio file recorder takes the frames from the file recorder and
calibrates them, and optionally wraps them as time data.