Implements a framework for processing classes
We are using the convention that certain classes can act both as data
containers and data processors. This is facilitated by a new metaclass
that modify how these classes are instantiated.
If instantiated with data that should be processed, first a "processor"
will be created with just the processing settings, then this processor
will be called and the output is returned.