Keep attributes in dB function
This will keep the xarray attributes when converting to dB. The only
attribute we use at the moment is the samplerate, stored in the time
coordinate.
Considering https://github.com/pydata/xarray/issues/8047, we need to
have keep_attrs=True at the moment, since we otherwise loose the attr
on the input array. I don't understand how it took this long for us
to realize that we strip samplerates on inputs when converting to dB...