Improves plotting
This moves much of the common stuff related to plotting to one place,
instead of having it spread out in lots of make_figure functions.
Now, most subclasses only have to extend a _figure_layout function
to add more layout styling to a figure.
A few cases, like Position, has other requirements on how a figure
is created. They can then re-implement the make_figure function, which
is just a few lines of code.