Improved concatenation of coordinates
This gives a better concatenation of multiple coordinate segments
when using multiple parts and placing nan-values between each part.
The main difference is that the concatenated coordinates will be
sequential, by repeating the last coordinate in each part once.
This is also extended with optional sorting of the values, which is
convenient whan reading multiple data files.