New method to implement calculation classes
This makes it easier to implement processing classes without having
to think too much about when a superclass can be called and when it
cannot. The "public" class that replaces the calculation class will
have both the methods from the data class and the calculation class,
but when instantiating the public class, one of the two specialized
classes will always be chosen.