Saves timedata with cosistent foramtting
xarray tries to use the least precise format that it can get away with
when storing time data. This breaks when iterating over times and
storing with only one time per save, since the first time will use
"days" as the resolution. We have numpy datetimes with nanosecond
resolution for all our timestamps, so we can always represent the times
with nanoseconds.