Replaces select_wrapper with from_dataset
The idea behind these two classmethods is almost the same - to select
the approproate wrapper for some dataset. The advantage of the from_dataset
method is that it also instantiates the new object, which allows it to
also modify the data if needed, e.g. squeeze surplus dims.
Since we now use this in the .sel and .isel methods, it also makes sure
that selecting a sensor from a sensor array will return a proper sensor
object.