Move filtering implementations
This moves the responsibility for filtering to a dedicated filterbank
class, instead of a pseudo-shared spectrogram/filterbank class.
This removes the need for complicated class decorators, in favor of a
few additional class methods to perform the analysis.
The classes used to interact with spectrograms are now moved to the
analysis namespace, which will likely be one of the main namespaces
that a user would interact with.