Improves rolling over large datasets on disk
When rolling over large datasets, they will be stored to disk. This
means we cannot load everything as a numpy array and just loop over that.
Looping and indexing the xarray object is too slow, since the xarray
methods has loads of overhead.
This changes the implementation to load chunks of the data using xarray
indexing, then loop over those chunks of numpy data.