Adds implementation to compute spectral probability from spectrogram
If a spectrogram is pre-computed, e.g., saved to disk, we don't want to
re-compute the spectrograms from the timedata just to compute the
spectral probability. This implementation will roll over the spectra in
the spectrogram just as if they were computed from the filterbank, and
apply the same averaging and binning as before.