Fixes infinite recursion on attribute lookup for temp objects
Some autocomplete engines construct temporary objects and try to get
attributes on them to provide better information. This can trigger lookup
of the _data attribute before the object is properly initialized. The
_data attribute will then not exist, which triggers the __getattr__
mehtod for lookup. Since we use self._data in this method, we end up
with infinite recursion and the kernel crashes. By directly specifying
to use __getattribute__ from object to get _data, we won't end up
re-entering in __getattr__ to get _data.